Move RequestAuthenticator and before() middleware to AuthServiceProvider
